Refactor transfers into transferList
